Rename some pthread tests from .cpp to .c. NFC
Also apply some reformatting to some.
