+++ test/pthread/test_pthread_create_pthread.c
@@ -3,6 +3,8 @@
 // University of Illinois/NCSA Open Source License.  Both these licenses can be
 // found in the LICENSE file.
 
+#define _GNU_SOURCE
+
 #include <pthread.h>
 #include <emscripten/console.h>
 #include <assert.h>
@@ -11,15 +13,13 @@
 
 volatile int result = 0;
 
-static void *thread2_start(void *arg)
-{
+static void *thread2_start(void *arg) {
   emscripten_out("thread2_start!");
   ++result;
   return NULL;
 }
 
-static void *thread1_start(void *arg)
-{
+static void *thread1_start(void *arg) {
   emscripten_out("thread1_start!");
   pthread_t thr;
   int rtn = pthread_create(&thr, NULL, thread2_start, NULL);
@@ -33,8 +33,7 @@ static void *thread1_start(void *arg)
 }
 
 #define DEFAULT_STACK_SIZE (64*1024)
-int main()
-{
+int main() {
   pthread_t thr;
   pthread_create(&thr, NULL, thread1_start, NULL);
 
@@ -44,8 +43,7 @@ int main()
   void *stack_addr;
   pthread_attr_getstack(&attr, &stack_addr, &stack_size);
   printf("stack_size: %d, stack_addr: %p\n", (int)stack_size, stack_addr);
-  if (stack_size != DEFAULT_STACK_SIZE || stack_addr == NULL)
-    result = -100; // Report failure.
+  assert(stack_size == DEFAULT_STACK_SIZE && stack_addr != NULL);
 
   pthread_join(thr, NULL);
