+++ test/pthread/test_pthread_nested_spawns.c
@@ -3,7 +3,7 @@
 // University of Illinois/NCSA Open Source License.  Both these licenses can be
 // found in the LICENSE file.
 
-#include <cassert>
+#include <assert.h>
 #include <stdio.h>
 #include <pthread.h>
